// apps/web/src/lib/api/generated/models/appraisalDefinitionDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */
import type { AppraisalApplicabilityDto } from './appraisalApplicabilityDto.ts';
import type { AppraisalDomainDto } from './appraisalDomainDto.ts';
import type { AppraisalJudgmentSchemaDto } from './appraisalJudgmentSchemaDto.ts';

export interface AppraisalDefinitionDto {
	applicability: AppraisalApplicabilityDto;
	description: string;
	domains: AppraisalDomainDto[];
	id: string;
	name: string;
	overall_judgment: AppraisalJudgmentSchemaDto;
	/** @minimum 0 */
	version: number;
}